OAuth2: Sends the user to a autofill registration form with the information recived from google. OAuth2 will also check if the user frist and last name exists, this messure will only be present during "fadderuka" to prevent a user from registering all his/her accounts to our site.
Registration: A Controller that handles autofill and submit for the registration form, it also controlls the terms agreement route

User(Model): Added constraints for each attribute, this is neceserry for the registration form. Removed name attribute and added first and last name instead. A new method added(genderAsList()) so that register form can access the Enum values as a list.

Form: A registration form that shows the user his/her information before registering them in the db. They also have to read and check of terms and agreement(brukeravtalen), this allows us to use their information as we want.

Terms: A view that describes "brukeravtalen". This could also be included in the registration form if preffered.

Routes: added routes for Registration.submit and Registration.terms, note that autofill is called without routing but with direct calls to the Register Controller. The idea is that clients should not have access to this Result.